+++ loritta-discord/src/main/java/com/mrpowergamerbr/loritta/tables/Profiles.kt
@@ -9,7 +9,7 @@ object Profiles : SnowflakeTable() {
 	val lastMessageSentAt = long("last_message_sent_at")
 	val lastMessageSentHash = integer("last_message_sent_hash")
 	val lastCommandSentAt = long("last_command_sent_at").nullable()
-	val money = double("money").index()
+	val money = long("money").index()
 	var isDonator = bool("donator")
 	var donatorPaid = double("donator_paid")
 	var donatedAt = long("donated_at")
